+++ Gruntfile.js
@@ -2,7 +2,7 @@ module.exports = function(grunt) {
   grunt.loadNpmTasks('grunt-sass');
   grunt.loadNpmTasks('grunt-contrib-cssmin');
   grunt.loadNpmTasks('grunt-contrib-copy');
-  grunt.loadNpmTasks('grunt-contrib-uglify');
+  // grunt.loadNpmTasks('grunt-contrib-uglify');
   grunt.loadNpmTasks('grunt-eslint');
   grunt.loadNpmTasks('grunt-contrib-watch');
   grunt.loadNpmTasks('grunt-protractor-runner');
@@ -45,21 +45,18 @@ module.exports = function(grunt) {
         }
       }
     },
-    uglify: {
-      options: {
-        sourceMap: true,
-        output: {
-          comments: 'some'
-        }
-      },
-      dist: {
-        files: {
-          'dist/jq/jquery.js': 'src/jq/jquery.js',
-          'dist/jq/jquery-ui.js': 'src/jq/jquery-ui.js',
-          'dist/jq/jquery.ui.touch-punch.js': 'src/jq/jquery.ui.touch-punch.js',
-        }
-      }
-    },
+    // uglify: {
+    //   options: {
+    //     sourceMap: true,
+    //     output: {
+    //       comments: 'some'
+    //     }
+    //   },
+    //   dist: {
+    //     files: {
+    //     }
+    //   }
+    // },
     eslint: {
       target: ['*.js', 'src/*.js']
     },
@@ -107,6 +104,6 @@ module.exports = function(grunt) {
   });
 
   grunt.registerTask('lint', ['eslint']);
-  grunt.registerTask('default', ['sass', 'cssmin', /*'eslint',*/ 'copy', 'uglify']);
+  grunt.registerTask('default', ['sass', 'cssmin', /*'eslint',*/ 'copy', /*'uglify'*/]);
   grunt.registerTask('e2e-test', ['connect', 'protractor_webdriver', 'protractor']);
 };